pyproject.toml
[tool.poetry]
name = "nd608-project-personalized-real-estate-agent"
version = "0.1.0"
description = ""
authors = ["Agustin Barto <abarto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
openai = "^1.23.2"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
lancedb = "^0.6.8"
pillow = "^10.3.0"
requests = "^2.31.0"
transformers = "^4.40.0"
torch = "^2.2.2"
gradio = "^4.28.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
